Split out LearningObjectivesTable
